+++ src/serious_python_android/lib/serious_python_android.dart
@@ -8,6 +8,7 @@ import 'package:path/path.dart' as p;
 import 'package:serious_python_platform_interface/serious_python_platform_interface.dart';
 
 import 'src/cpython.dart';
+import 'src/log.dart';
 
 /// An implementation of [SeriousPythonPlatform] that uses method channels.
 class SeriousPythonAndroid extends SeriousPythonPlatform {
@@ -33,10 +34,9 @@ class SeriousPythonAndroid extends SeriousPythonPlatform {
       List<String>? modulePaths,
       Map<String, String>? environmentVariables,
       bool? sync}) async {
-    Future setenv(String key, String value) async {
-      await methodChannel.invokeMethod<String>(
-          'setEnvironmentVariable', {'name': key, 'value': value});
-    }
+    Future<void> setenv(String key, String value) =>
+        methodChannel.invokeMethod<String>(
+            'setEnvironmentVariable', {'name': key, 'value': value});
 
     // load libpyjni.so to get JNI reference
     try {
@@ -44,13 +44,13 @@ class SeriousPythonAndroid extends SeriousPythonPlatform {
           .invokeMethod<String>('loadLibrary', {'libname': 'pyjni'});
       await setenv("FLET_JNI_READY", "1");
     } catch (e) {
-      debugPrint("Warning: Unable to load libpyjni.so library: $e");
+      spDebug("Unable to load libpyjni.so library: $e");
     }
 
     // unpack python bundle
     final nativeLibraryDir =
         await methodChannel.invokeMethod<String>('getNativeLibraryDir');
-    debugPrint("getNativeLibraryDir: $nativeLibraryDir");
+    spDebug("getNativeLibraryDir: $nativeLibraryDir");
 
     var bundlePath = "$nativeLibraryDir/libpythonbundle.so";
     var sitePackagesZipPath = "$nativeLibraryDir/libpythonsitepackages.so";
@@ -60,7 +60,7 @@ class SeriousPythonAndroid extends SeriousPythonPlatform {
     }
     var pythonLibPath =
         await extractFileZip(bundlePath, targetPath: "python_bundle");
-    debugPrint("pythonLibPath: $pythonLibPath");
+    spDebug("pythonLibPath: $pythonLibPath");
 
     var programDirPath = p.dirname(appPath);
 
@@ -74,22 +74,22 @@ class SeriousPythonAndroid extends SeriousPythonPlatform {
     if (await File(sitePackagesZipPath).exists()) {
       var sitePackagesPath = await extractFileZip(sitePackagesZipPath,
           targetPath: "python_site_packages");
-      debugPrint("sitePackagesPath: $sitePackagesPath");
+      spDebug("sitePackagesPath: $sitePackagesPath");
       moduleSearchPaths.add(sitePackagesPath);
     }
 
-    setenv("PYTHONINSPECT", "1");
-    setenv("PYTHONDONTWRITEBYTECODE", "1");
-    setenv("PYTHONNOUSERSITE", "1");
-    setenv("PYTHONUNBUFFERED", "1");
-    setenv("LC_CTYPE", "UTF-8");
-    setenv("PYTHONHOME", pythonLibPath);
-    setenv("PYTHONPATH", moduleSearchPaths.join(":"));
+    await setenv("PYTHONINSPECT", "1");
+    await setenv("PYTHONDONTWRITEBYTECODE", "1");
+    await setenv("PYTHONNOUSERSITE", "1");
+    await setenv("PYTHONUNBUFFERED", "1");
+    await setenv("LC_CTYPE", "UTF-8");
+    await setenv("PYTHONHOME", pythonLibPath);
+    await setenv("PYTHONPATH", moduleSearchPaths.join(":"));
 
     // set environment variables
     if (environmentVariables != null) {
       for (var v in environmentVariables.entries) {
-        setenv(v.key, v.value);
+        await setenv(v.key, v.value);
       }
     }