+++ src/serious_python_android/lib/src/log.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/foundation.dart';
+
+void spDebug(String message) {
+  if (message.startsWith('[serious_python]')) {
+    debugPrint(message);
+  } else {
+    debugPrint('[serious_python] $message');
+  }
+}
+